[feature] implement multithreaded architecture
Before this commit, the application was single threaded.
In particular, the data-getting functionality worked by having a simple timer
fire on a predetermined interval, which blocked the event (aka UI) loop.
That's not good.

This commit implements a multithreaded architecture, where the event loop runs
in its own thread, and all the data getting stuff has another thread all to
itself.
Currently the program just generates random numbers, but adding in Bluetooth
support should be more-or-less drop-in (famous last words).

The way it works is that the data thread emits a newDataAvailable signal every
time the it generates a new value.
The UI thread has a onNewDataAvailable slot that's waiting for a signal, and
upon receipt it fires.
The slot takes those new data values it got from the data thread and updates the
UI with them.
This repeats ad-infinitum.