[feature] add logging functionality
Seeing data in real time is nice and all, but keeping it around for later
analysis is important.
This change adds such functionality, creating a new logfile for each run of the
application and sending data values to it each time new data is received.
This feature runs in the data thread, so it does not interfere with the event
loop.