[fix] kill all threads when application is closed
Prior to this change, the data thread would keep running in the background even
though the user had closed the UI, so you had to kill it in task manager.

This change adds an signal in MainWindow that's emitted when the user closes the
window.
This signal is picked up by a slot in the genBoardInterface signal, which flips
a bool to end the thread if the signal is received.
The worker thread is constantly checking that bool, so as soon as it gets it,
the thread closes and everything is fine.