[feature] add max values column to UI
Remembering values in your head during the test is hard with so much other stuff
going on.
With this change, the UI keeps track of the largest values itself and presents
them clearly to the user.

The way I implemented this uses arrays, which means that instead of using a
bunch of private variables in genBoardInterface, the main data transfer format
is an array.
I'll change the rest of the program to use arrays in the next commit, but for
now it doesn't.
As a result the plots don't actually display anything.