[feature] add highlighting of values beyond thresholds in dataTable
With this change, the user can define a set of thresholds for each data value.
When these "warning thresholds" are exceeded, the UI will make the table cell
containing these values yellow to highlight to the user that something is up.

As an example, one could define a threshold for the engine temp as 200 degrees,
and if the temp of the engine exceeded that value the UI would make the engine
temp cell yellow.

An extension to this I intend to add later would be to add a second, more urgent
set of user-definable thresholds which would cause immediate shutoff of the
engine if they were exceeded.