[feature] Add typical minimum/maximum config option
This'll be used for the upcoming revamp of the plotting
feature.

The typical values in the config.toml file are representative
of the weird values that come off BT during benchtop
testing.
They will have to change once we get into actual testing.

Also includes a clang-format run and some changed formatting
for readablility.