[feature] add scaledValue [0,100] for plots
The plots will by default plot scaled values that will be mapped
from whatever a typical min/max is to a range of [0,100].

This is because it isn't practical to have all of the data coming
off BT on the same scale.

If the user wants to see the actual unscaled value, they'll be able
to toggle a checkbox or radio button to have the second y-axis that'll
be unscaled show up.

That's forthcoming in the next few commits.